Fix bug with SVS paths/files containing "."
When constructing the output filename, a string substitution was used that replaced everything after "." with ".jpg"; however, because that is a greedy match, it would replace anything after the first "." in the full file path with "jpg".  The new statement uses the "%" operator to delete the shortest match to ".*" from the back of the string and then adds ".jpg" to construct the output filename.