Fixed silly error in extract_SVS_slide_labels.sh
Somehow I never noticed that there was a copy/paste error that prevented the script from correctly parsing the input path argument; as a result, the script would always default to using the cwd as the input path.  Apparently this script had only been run from the same directory as what was specified in the input path argument, and so the error was never noticed.